+++ web/app/controllers/authenticated/all.ts
@@ -1,8 +1,4 @@
 import Controller from "@ember/controller";
-import { SortByValue } from "hermes/components/header/toolbar";
-import { SortDirection } from "hermes/components/table/sortable-header";
-import AuthenticatedAllRoute from "hermes/routes/authenticated/all";
-import { ModelFrom } from "hermes/types/route-models";
 
 export default class AuthenticatedAllController extends Controller {
   queryParams = ["docType", "owners", "page", "product", "sortBy", "status"];
@@ -12,15 +8,4 @@ export default class AuthenticatedAllController extends Controller {
   product = [];
   sortBy = "dateDesc";
   status = [];
-
-  declare model: ModelFrom<AuthenticatedAllRoute>;
-
-  get sortDirection() {
-    switch (this.model.sortedBy) {
-      case SortByValue.DateAsc:
-        return SortDirection.Asc;
-      default:
-        return SortDirection.Desc;
-    }
-  }
 }
