+++ web/app/controllers/authenticated/documents.ts
@@ -1,4 +1,8 @@
 import Controller from "@ember/controller";
+import { SortByValue } from "hermes/components/header/toolbar";
+import { SortDirection } from "hermes/components/table/sortable-header";
+import AuthenticatedDocumentsRoute from "hermes/routes/authenticated/documents";
+import { ModelFrom } from "hermes/types/route-models";
 
 export default class AuthenticatedDocumentsController extends Controller {
   queryParams = ["docType", "owners", "page", "product", "sortBy", "status"];
@@ -8,4 +12,15 @@ export default class AuthenticatedDocumentsController extends Controller {
   product = [];
   sortBy = "dateDesc";
   status = [];
+
+  declare model: ModelFrom<AuthenticatedDocumentsRoute>;
+
+  get sortDirection() {
+    switch (this.model.sortedBy) {
+      case SortByValue.DateAsc:
+        return SortDirection.Asc;
+      default:
+        return SortDirection.Desc;
+    }
+  }
 }
