+++ web/app/routes/authenticated/documents.ts
@@ -4,6 +4,7 @@ import ConfigService from "hermes/services/config";
 import AlgoliaService from "hermes/services/algolia";
 import { DocumentsRouteParams } from "hermes/types/document-routes";
 import ActiveFiltersService from "hermes/services/active-filters";
+import { SortByValue } from "hermes/components/header/toolbar";
 
 export default class AuthenticatedDocumentsRoute extends Route {
   @service("config") declare configSvc: ConfigService;
@@ -32,6 +33,8 @@ export default class AuthenticatedDocumentsRoute extends Route {
   };
 
   async model(params: DocumentsRouteParams) {
+    const sortedBy = params.sortBy ?? SortByValue.DateDesc;
+
     const searchIndex =
       params.sortBy === "dateAsc"
         ? this.configSvc.config.algolia_docs_index_name + "_createdTime_asc"
@@ -44,6 +47,6 @@ export default class AuthenticatedDocumentsRoute extends Route {
 
     this.activeFilters.update(params);
 
-    return { facets, results };
+    return { facets, results, sortedBy };
   }
 }
